feat!: Replace `SequenceNumber` with SDK `Version` (#9794)
Fixes #9835
Replaces the `SequenceNumber` in the `iota-types` crate with the
`Version` from the SDK.

- [ ] Protocol:
- [ ] Nodes (Validators and Full nodes):
- [ ] Indexer:
- [ ] JSON-RPC:
- [ ] GraphQL:
- [ ] CLI:
- [x] Rust SDK: Some methods have been removed from SequenceNumber due
to crate externalization (iota-rust-sdk::Version is replacing
iota-types::SequenceNumber).
- [ ] REST API:

+++ crates/iota-json-rpc-types/src/iota_transaction.rs
@@ -1033,19 +1033,14 @@ impl<T: TransactionEffectsAPI> From<T> for IotaTransactionBlockEffectsV1 {
 fn owned_objref_string(obj: &OwnedObjectRef) -> String {
     format!(
         " ┌──\n │ ID: {} \n │ Owner: {} \n │ Version: {} \n │ Digest: {}\n └──",
-        obj.reference.object_id,
-        obj.owner,
-        u64::from(obj.reference.version),
-        obj.reference.digest
+        obj.reference.object_id, obj.owner, obj.reference.version, obj.reference.digest
     )
 }
 
 fn objref_string(obj: &IotaObjectRef) -> String {
     format!(
         " ┌──\n │ ID: {} \n │ Version: {} \n │ Digest: {}\n └──",
-        obj.object_id,
-        u64::from(obj.version),
-        obj.digest
+        obj.object_id, obj.version, obj.digest
     )
 }
